Feature/artwork - 팔로우 및 좋아요 버튼

See merge request s06-webmobile2-sub2/S06P12B204!75

+++ frontend/src/components/artwork/child/CommentItem.vue
@@ -94,6 +94,7 @@ export default defineComponent({
       });
     },
     deleteComment() {
+      if (!confirm("정말 삭제하시겠습니까?")) return;
       CommentsAPI.deleteComment(this.comment.id).then((res: ResponseData) => {
         this.$emit("getList");
         console.log(res);
